requirements.txt
streamlit==1.53.1
mediapipe==0.10.30
opencv-python-headless==4.8.1.78
numpy==1.26.4
pillow==12.1.0

